handle here doc cases

+++ srcs/execution.c
@@ -6,7 +6,7 @@
 /*   By: nqasem <nqasem@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/09 19:55:32 by nqasem            #+#    #+#             */
-/*   Updated: 2025/05/05 20:26:52 by nqasem           ###   ########.fr       */
+/*   Updated: 2025/05/06 21:06:56 by nqasem           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -72,8 +72,10 @@ int searching_here_doc(t_cmd **cmd, t_here_doc **here_doc)
 	int		check_error;
 	char	*temp;
 	int		i;
+	int		p;
 
 	i = -1;
+	p = 0;
 	current = (*cmd)->word;
 	(*here_doc)->counter = 0;
 	while (-1 < ++i && current != NULL)
@@ -89,11 +91,12 @@ int searching_here_doc(t_cmd **cmd, t_here_doc **here_doc)
 				(*cmd)->exit_status = -14;
 				return (-1);
 			}
-			(*here_doc)->pryority[i] = 1;
+			(*here_doc)->pryority[i] = 2 + p;
 			(*here_doc)->counter++;
+			p++;
 		}
 		else
-			(*here_doc)->pryority[i] = 2;
+			(*here_doc)->pryority[i] = 1;
 		current = current->next;
 	}
  		(*here_doc)->pryority[i] = '\0';
@@ -120,7 +123,7 @@ int		here_doc_manger(t_cmd **cmd, char **file_loc)
 	current = (*cmd)->word;
 	while ((*cmd)->here_doc->pryority[i] != '\0')
 	{
-		if ((*cmd)->here_doc->pryority[i] == 1)
+		if ((*cmd)->here_doc->pryority[i] >= 2)
 		{
 				if (heredoc(current->content, &(file_loc[j])))
 				{
@@ -135,12 +138,7 @@ int		here_doc_manger(t_cmd **cmd, char **file_loc)
 
 	return (0);
 }
- 	// int p = 0;
-	// while ((*cmd)->here_doc->pryority[p] != '\0')
-	// {
-	// 	printf("Command: %d\n", (*cmd)->here_doc->pryority[p]);
-	// 	p++;
-	// }
+
 int	execution(t_cmd **cmd, char **env)
 {
 	int		i;
@@ -148,6 +146,7 @@ int	execution(t_cmd **cmd, char **env)
 	int		k;
 	int		size;
 	t_list	*current;
+	char **file_loc;
 
 	size = ft_lstsize((*cmd)->word);
 	(*cmd)->here_doc->pryority = malloc(sizeof(int) * (size + 1));
@@ -159,16 +158,19 @@ int	execution(t_cmd **cmd, char **env)
 			perror("Error in here doc");
 			return (-13);
  	}
-	char **file_loc = malloc(sizeof(char *) * ((*cmd)->here_doc->counter + 1));
-	if (!file_loc)
+	file_loc = NULL;
+	if ((*cmd)->here_doc->counter > 0)
 	{
-		perror("malloc");
-		return (-1);
+		file_loc = malloc(sizeof(char *) * ((*cmd)->here_doc->counter + 1));
+		if (!file_loc)
+		{
+			perror("malloc");
+			return (-1);
+		}
+		(*cmd)->here_doc->file_loc = file_loc;
+		if(here_doc_manger(cmd, (*cmd)->here_doc->file_loc))
+			return (-1);
 	}
-	(*cmd)->here_doc->file_loc = file_loc;
-	if(here_doc_manger(cmd, (*cmd)->here_doc->file_loc))
-		return (-1);
-
 	pid_t	pids[size];
 	int		pipe_fd2[size][2];
 	current = (*cmd)->word;
@@ -194,18 +196,25 @@ int	execution(t_cmd **cmd, char **env)
 		}
 		i++;
 	}
-	// if(child_process(cmd, &current, pipe_fd2, (pids), file_loc))
 	if(child_process(cmd, &current, pipe_fd2,  pids, file_loc))
 		return (-1);
-	int no = 0;
-	while (no < (*cmd)->here_doc->counter)
-	{
-		unlink((*cmd)->here_doc->file_loc[no]);
-		free((*cmd)->here_doc->file_loc[no]);
-		no++;
-	}
 	close_wait(pids, size, pipe_fd2);
- 	return (0);
+    int no = 0;
+    while (no < (*cmd)->here_doc->counter)
+    {
+        if ((*cmd)->here_doc->file_loc[no]) 
+		{
+            unlink((*cmd)->here_doc->file_loc[no]);
+            free((*cmd)->here_doc->file_loc[no]);
+        }
+        no++;
+    }
+	if (file_loc)
+	{
+    	free((*cmd)->here_doc->file_loc);
+    	(*cmd)->here_doc->file_loc = NULL;
+	}
+	return (0);
 }
 
 int		child_process(t_cmd **cmd, t_list	**current,  int pipe_fd2[][2], pid_t pids[], char **file_loc)
@@ -229,7 +238,6 @@ int		child_process(t_cmd **cmd, t_list	**current,  int pipe_fd2[][2], pid_t pids
 				return (-1);
 			if(dup_process_2(cmd, (current), file_loc, i) == -1)
 				return (-1);
-				dprintf(2, "print: %d\n", (*cmd)->here_doc->index);
 		}
 		if (i > 0)
 			close(pipe_fd2[i - 1][0]);
@@ -244,38 +252,53 @@ int		child_process(t_cmd **cmd, t_list	**current,  int pipe_fd2[][2], pid_t pids
 int		execute_heredoc(char *file, char **ev, int i, char **file_loc)
 {
     if (!file_loc[i])
-	{
+    {
         dprintf(2, "Error: file_loc[%d] is NULL in execute_heredoc\n", i);
         return (-1);
     }
     int fd = open(file_loc[i], O_RDONLY);
     if (fd < 0)
-	{
+    {
         perror("open heredoc file");
         return (-1);
     }
-    int org = dup(STDIN_FILENO);
     dup2(fd, STDIN_FILENO);
-	char *temp = ft_strfchr(file, '<');
-	if (ft_strlen(temp) <= 1)
-		return (0);
+    char *temp = ft_strfchr(file, '<');
+    close(fd);
+    if (ft_strlen(temp) <= 1)
+	{
+        free(temp);
+        return (-1);
+    }
     ft_execve(temp, ev);
+    free(temp);
     return (0);
 }
 
-int		dup_process_2(t_cmd **cmd, t_list	**current, char **file_loc, int i)
+int		dup_process_2(t_cmd **cmd, t_list **current, char **file_loc, int i)
 {
 	char	**redirection_split;
 	char	**env;
 	int		heredoc_idx;
 
 	env = (*cmd)->env;
-	if ((*cmd)->here_doc->pryority[i] == 1)
+	if ((*cmd)->here_doc->pryority[i] >= 2)
 	{
-        // heredoc_idx = (*cmd)->here_doc->index;
-        // if (execute_heredoc((*current)->content, env, heredoc_idx, file_loc) == -1)
-        //     return (-1);
-  
+		write(1,"abdullah and cut\n",18);
+        heredoc_idx = (*cmd)->here_doc->pryority[i] - 2;
+        if (execute_heredoc((*current)->content, env, heredoc_idx, file_loc) == -1)
+		{
+			int no = 0;
+			while (no < (*cmd)->here_doc->counter)
+			{
+				if ((*cmd)->here_doc->file_loc[no]) 
+				{
+					free((*cmd)->here_doc->file_loc[no]);
+				}
+				no++;
+			}
+            return (-1);
+		}
 	}
 	if (ft_strmchr((*current)->content ,"<>") && (*cmd)->who_am_i != 13)
 	{
@@ -284,11 +307,6 @@ int		dup_process_2(t_cmd **cmd, t_list	**current, char **file_loc, int i)
 			write(2, "Error: Invalid redirection\n\n", 27);
 			return (-1);
 		}
-	}
-	if ((*cmd)->who_am_i == 13)
-	{
-		write(2, "here doc\n", 9);
-		return (-1);
 	}
 	if (ft_execve((*current)->content, env) == -1)
 	{
